move getRemoteServer to ODistributed plugin
